Изменение : обновление jsoncpp 1.7.5
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@1869 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/ExtLib/jsoncpp/include/json/config.h
@@ -80,7 +80,7 @@
 // In c++11 the override keyword allows you to explicity define that a function
 // is intended to override the base-class version.  This makes the code more
 // managable and fixes a set of common hard-to-find bugs.
-#if __cplusplus >= 201103L 
+#if __cplusplus >= 201103L
 # define JSONCPP_OVERRIDE override
 #elif defined(_MSC_VER) && _MSC_VER > 1600
 # define JSONCPP_OVERRIDE override
@@ -152,8 +152,8 @@ typedef unsigned int LargestUInt;
 typedef __int64 Int64;
 typedef unsigned __int64 UInt64;
 #else                 // if defined(_MSC_VER) // Other platforms, use long long
-typedef long long int Int64;
-typedef unsigned long long int UInt64;
+typedef int64_t Int64;
+typedef uint64_t UInt64;
 #endif // if defined(_MSC_VER)
 typedef Int64 LargestInt;
 typedef UInt64 LargestUInt;
